调整 LoadSource 和 Compile 方法以提高代码可读性
对 LoadSource 方法中的 CreateBlob 参数进行类型转换，将 includeBytes.Length 从 int 转换为 uint。对 Compile 方法的签名进行格式化调整，将参数分成多行以提高可读性。

+++ Source/Graphics/Graphics.Vulkan/Helpers/DxcHelpers.cs
@@ -13,13 +13,18 @@ public static unsafe class DxcHelpers
         {
             byte[] includeBytes = includeResolver?.Invoke(filename) ?? [];
 
-            includeSource = DxcCompiler.Utils.CreateBlob((nint)includeBytes.AsPointer(), includeBytes.Length, Dxc.DXC_CP_UTF8);
+            includeSource = DxcCompiler.Utils.CreateBlob((nint)includeBytes.AsPointer(),
+                                                         (uint)includeBytes.Length,
+                                                         Dxc.DXC_CP_UTF8);
 
             return Result.Ok;
         }
     }
 
-    public static byte[] Compile(ShaderStages stage, string hlsl, string entryPoint, Func<string, byte[]>? includeResolver = null)
+    public static byte[] Compile(ShaderStages stage,
+                                 string hlsl,
+                                 string entryPoint,
+                                 Func<string, byte[]>? includeResolver = null)
     {
         using IncludeHandler includeHandler = new(includeResolver);
